+++ app/src/main/java/com/example/android/digiteen/Adapter/TabAdapter.java
@@ -1,10 +1,8 @@
-package com.example.android.digiteen;
+package com.example.android.digiteen.Adapter;
 
-import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentPagerAdapter;
-import androidx.viewpager2.adapter.FragmentStateAdapter;
 
 import java.util.List;
