+++ app/src/main/java/com/example/android/digiteen/View/Student/StudentSignUpFragment.java
@@ -1,4 +1,4 @@
-package com.example.android.digiteen;
+package com.example.android.digiteen.View.Student;
 
 import android.app.ProgressDialog;
 import android.os.Build;
@@ -9,7 +9,6 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.ProgressBar;
 import android.widget.Toast;
 
 import androidx.annotation.NonNull;
@@ -20,6 +19,7 @@ import androidx.navigation.NavController;
 import androidx.navigation.NavOptions;
 import androidx.navigation.Navigation;
 
+import com.example.android.digiteen.R;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
