+++ app/src/main/res/layout/fragment_login.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:fitsSystemWindows="true"
-    tools:context=".LoginFragment">
+    tools:context=".View.LoginFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
