+++ app/src/main/res/layout/fragment_owner_landing.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".OwnerLandingFragment">
+    tools:context=".View.Owner.OwnerLandingFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
