+++ app/src/main/res/layout/fragment_student_landing.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".StudentLandingFragment">
+    tools:context=".View.Student.StudentLandingFragment">
 
     <Button
         android:id="@+id/my_orders"
