+++ app/src/main/res/layout/fragment_student_sign_up.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".StudentSignUpFragment">
+    tools:context=".View.Student.StudentSignUpFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
